Added 2 new endpoints: 	- Adding a workday at weekend 	- Getting the statistics of a given month
All the enpoints now return a Response object, if one of the user-defined exceptions
are thrown, an HTML code specific to this Exception is sent in the object.

Added new test cases to the SoapUI tests based on the modifications and the new endpoint

Tasks can now be created with time formats: "HH:MM", "HHMM", "H:MM", "HMM".